--- instrumentation/grizzly-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/grizzly/DefaultFilterChainAdvice.java
@@ -1,24 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.grizzly;
-
-import static io.opentelemetry.javaagent.instrumentation.grizzly.GrizzlyHttpServerTracer.tracer;
-
-import io.opentelemetry.context.Context;
-import net.bytebuddy.asm.Advice;
-import org.glassfish.grizzly.filterchain.FilterChainContext;
-
-public class DefaultFilterChainAdvice {
-
-  @Advice.OnMethodEnter(suppress = Throwable.class)
-  public static void onFail(
-      @Advice.Argument(0) FilterChainContext ctx, @Advice.Argument(1) Throwable throwable) {
-    Context context = tracer().getServerContext(ctx);
-    if (context != null) {
-      tracer().endExceptionally(context, throwable);
-    }
-  }
-}
