+++ instrumentation/grizzly-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/grizzly/DefaultFilterChainInstrumentation.java
@@ -5,17 +5,21 @@
 
 package io.opentelemetry.javaagent.instrumentation.grizzly;
 
+import static io.opentelemetry.javaagent.instrumentation.grizzly.GrizzlyHttpServerTracer.tracer;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPrivate;
 import static net.bytebuddy.matcher.ElementMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
+import io.opentelemetry.context.Context;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import java.util.Collections;
 import java.util.Map;
+import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
+import org.glassfish.grizzly.filterchain.FilterChainContext;
 
 public class DefaultFilterChainInstrumentation implements TypeInstrumentation {
 
@@ -32,6 +36,18 @@ public class DefaultFilterChainInstrumentation implements TypeInstrumentation {
             .and(named("notifyFailure"))
             .and(takesArgument(0, named("org.glassfish.grizzly.filterchain.FilterChainContext")))
             .and(takesArgument(1, Throwable.class)),
-        DefaultFilterChainAdvice.class.getName());
+        DefaultFilterChainInstrumentation.class.getName() + "$NotifyFailureAdvice");
+  }
+
+  public static class NotifyFailureAdvice {
+
+    @Advice.OnMethodEnter(suppress = Throwable.class)
+    public static void onFail(
+        @Advice.Argument(0) FilterChainContext ctx, @Advice.Argument(1) Throwable throwable) {
+      Context context = tracer().getServerContext(ctx);
+      if (context != null) {
+        tracer().endExceptionally(context, throwable);
+      }
+    }
   }
 }
