--- instrumentation/grizzly-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/grizzly/HttpCodecFilterOldAdvice.java
@@ -1,37 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.grizzly;
-
-import static io.opentelemetry.javaagent.instrumentation.grizzly.GrizzlyHttpServerTracer.tracer;
-
-import io.opentelemetry.context.Context;
-import java.lang.reflect.Method;
-import net.bytebuddy.asm.Advice;
-import org.glassfish.grizzly.filterchain.FilterChainContext;
-import org.glassfish.grizzly.http.HttpPacketParsing;
-import org.glassfish.grizzly.http.HttpRequestPacket;
-
-public class HttpCodecFilterOldAdvice {
-
-  @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-  public static void onExit(
-      @Advice.Origin Method method,
-      @Advice.Argument(0) FilterChainContext ctx,
-      @Advice.Argument(1) HttpPacketParsing httpHeader) {
-    Context context = tracer().getServerContext(ctx);
-
-    // only create a span if there isn't another one attached to the current ctx
-    // and if the httpHeader has been parsed into a HttpRequestPacket
-    if (context != null || !(httpHeader instanceof HttpRequestPacket)) {
-      return;
-    }
-    HttpRequestPacket httpRequest = (HttpRequestPacket) httpHeader;
-
-    // We don't want to attach the new context to this thread, as actual request will continue on
-    // some other thread where we will read and attach it via tracer().getServerContext(ctx).
-    tracer().startSpan(httpRequest, httpRequest, ctx, method);
-  }
-}
