--- instrumentation/grizzly-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/grizzly/HttpHandlerAdvice.java
@@ -1,21 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.grizzly;
-
-import static io.opentelemetry.javaagent.instrumentation.grizzly.GrizzlyHttpServerTracer.tracer;
-
-import io.opentelemetry.javaagent.instrumentation.api.Java8BytecodeBridge;
-import net.bytebuddy.asm.Advice;
-
-public class HttpHandlerAdvice {
-
-  @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-  public static void onExit(@Advice.Thrown Throwable throwable) {
-    if (throwable != null) {
-      tracer().onException(Java8BytecodeBridge.currentContext(), throwable);
-    }
-  }
-}
