+++ instrumentation/grizzly-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/grizzly/HttpServerFilterInstrumentation.java
@@ -5,16 +5,21 @@
 
 package io.opentelemetry.javaagent.instrumentation.grizzly;
 
+import static io.opentelemetry.javaagent.instrumentation.grizzly.GrizzlyHttpServerTracer.tracer;
 import static net.bytebuddy.matcher.ElementMatchers.isPrivate;
 import static net.bytebuddy.matcher.ElementMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
+import io.opentelemetry.context.Context;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import java.util.Collections;
 import java.util.Map;
+import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
+import org.glassfish.grizzly.filterchain.FilterChainContext;
+import org.glassfish.grizzly.http.HttpResponsePacket;
 
 public class HttpServerFilterInstrumentation implements TypeInstrumentation {
 
@@ -32,6 +37,19 @@ public class HttpServerFilterInstrumentation implements TypeInstrumentation {
             .and(takesArgument(2, named("org.glassfish.grizzly.http.HttpResponsePacket")))
             .and(takesArgument(3, named("org.glassfish.grizzly.http.HttpContent")))
             .and(isPrivate()),
-        HttpServerFilterAdvice.class.getName());
+        HttpServerFilterInstrumentation.class.getName() + "$PrepareResponseAdvice");
+  }
+
+  public static class PrepareResponseAdvice {
+
+    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
+    public static void onExit(
+        @Advice.Argument(0) FilterChainContext ctx,
+        @Advice.Argument(2) HttpResponsePacket response) {
+      Context context = tracer().getServerContext(ctx);
+      if (context != null) {
+        tracer().end(context, response);
+      }
+    }
   }
 }
